refactor(color-picker): removed unneeded index.ts files
