+++ src/demo-app/app/color-picker/color-picker.module.ts
@@ -9,12 +9,12 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatTabsModule } from '@angular/material/tabs';
 import { RouterModule } from '@angular/router';
-import { MccColorPickerModule } from '../../../lib/color-picker';
 import { ColorPickerComponent } from './color-picker.component';
 import { routes } from './color-picker.router';
 import { ColorPickerAlphaComponent } from './components/color-picker-alpha.component';
 import { ColorPickerApiComponent } from './components/color-picker-api.component';
 import { ColorPickerExamplesComponent } from './components/color-picker-examples.component';
+import { MccColorPickerModule } from '../../../lib/color-picker/public_api';
