+++ src/demo-app/app/color-picker/components/color-picker-alpha.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { EMPTY_COLOR, ENABLE_ALPHA_SELECTOR } from '../../../../lib/color-picker';
+import { EMPTY_COLOR, ENABLE_ALPHA_SELECTOR } from '../../../../lib/color-picker/public_api';
 
 @Component({
   selector: 'app-color-picker-alpha',
