+++ src/demo-app/app/scrollspy/scrollspy.module.ts
@@ -6,10 +6,10 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatTabsModule } from '@angular/material/tabs';
 import { RouterModule } from '@angular/router';
-import { MccScrollspyModule } from '../../../lib/scrollspy';
 import { ScrollspyExamplesComponent } from './components/scrollspy-examples.component';
 import { ScrollspyComponent } from './scrollspy.component';
 import { routes } from './scrollspy.router';
+import { MccScrollspyModule } from '../../../lib/scrollspy/public_api';
