+++ src/demo-app/app/speed-dial/speed-dial.module.ts
@@ -9,11 +9,11 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatRadioModule } from '@angular/material/radio';
 import { MatTabsModule } from '@angular/material/tabs';
 import { RouterModule } from '@angular/router';
-import { MccSpeedDialModule } from '../../../lib/speed-dial';
 import { SpeedDialApiComponent } from './components/speed-dial-api.component';
 import { SpeedDialExamplesComponent } from './components/speed-dial-examples.component';
 import { SpeedDialComponent } from './speed-dial.component';
 import { routes } from './speed-dial.router';
+import { MccSpeedDialModule } from '../../../lib/speed-dial/public_api';
 
 @NgModule({
   imports: [
