+++ src/lib/color-picker/public_api.ts
@@ -1,3 +1,12 @@
 export { MccColorPickerModule } from './color-picker.module';
 export { MccColorPickerService } from './color-picker.service';
-export { MccColorPickerItem, EMPTY_COLOR, ENABLE_ALPHA_SELECTOR } from './color-picker.types';
+export {
+  MccColorPickerItem,
+  EMPTY_COLOR,
+  ENABLE_ALPHA_SELECTOR,
+  COLOR_STRING_FORMAT,
+  DISABLE_SELECTED_COLOR_ICON,
+  SELECTED_COLOR_ICON,
+  SELECTED_COLOR_SVG_ICON,
+  USED_COLORS
+} from './color-picker.types';
