--- src/lib/speed-dial/index.ts
@@ -1,1 +1,0 @@
-export * from './public_api';
